+++ nodeconductor/structure/tests/serializers.py
@@ -8,18 +8,12 @@ from . import models
 class ServiceSerializer(structure_serializers.BaseServiceSerializer):
     class Meta(structure_serializers.BaseServiceSerializer.Meta):
         model = models.TestService
-        extra_kwargs = {
-            'url': {'lookup_field': 'uuid', 'view_name': 'test-detail'},
-            'customer': {'lookup_field': 'uuid'},
-            'settings': {'lookup_field': 'uuid'},
-        }
 
 
 class ServiceProjectLinkSerializer(structure_serializers.BaseServiceProjectLinkSerializer):
     class Meta(structure_serializers.BaseServiceProjectLinkSerializer.Meta):
         model = models.TestServiceProjectLink
         extra_kwargs = {
-            'url': {'view_name': 'test-spl-detail'},
             'service': {'lookup_field': 'uuid', 'view_name': 'oracle-detail'},
         }
 
@@ -38,9 +32,6 @@ class InstanceSerializer(structure_serializers.BaseResourceSerializer):
 
     class Meta(structure_serializers.BaseResourceSerializer.Meta):
         model = models.TestInstance
-        extra_kwargs = {
-            'url': {'lookup_field': 'uuid', 'view_name': 'test-instances-detail'},
-        }
 
 
 class NewInstanceSerializer(structure_serializers.BaseResourceSerializer):
@@ -57,6 +48,3 @@ class NewInstanceSerializer(structure_serializers.BaseResourceSerializer):
 
     class Meta(structure_serializers.BaseResourceSerializer.Meta):
         model = models.TestNewInstance
-        extra_kwargs = {
-            'url': {'view_name': 'test-new-instances-detail'}
-        }